requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
groq==0.4.1
resend==0.6.0
python-dotenv==1.0.0
lxml==4.9.3
httpx==0.27.2
httpcore==0.18.4
pandas==2.0.3
streamlit-shadcn-ui==0.1.6
apscheduler==3.10.4
youtube-transcript-api==1.2.3
google-api-python-client==2.108.0
tweepy==4.14.0
